+++ ChangeLog.md
@@ -11,6 +11,7 @@ Blob:
 - Added more allowed value to list blob request "include" query parameter:'tags', 'versions', 'deletedwithversions', 'immutabilitypolicy', 'legalhold', 'permissions'
 - Added more allowed value to list container request "include" query parameter: 'deleted'
 - Raise 416 when start range is bigger than blob length
+- Do not decode the rscd, rsce, rscl and rsct url parameters
 
 Queue: